refactor * add middleware to add user instance to request * change src structure * add `env` and `redis-3.2.11` directory to gitignore
