+++ .gitignore
@@ -67,3 +67,6 @@ fitness/media/
 
 #ansible
 *.retry
+
+env/
+redis-3.2.11/
