+++ src/main.py
@@ -1,27 +1,28 @@
 import asyncio
 
-import jinja2
 import aiohttp_jinja2
-
+import jinja2
 import uvloop
 from aiohttp import web
-from aioredis import create_pool
-from aiohttp_session import setup
+from aiohttp_session import session_middleware
 from aiohttp_session.redis_storage import RedisStorage
+from aioredis import create_pool
 
-from chat.config import setup_routes, settings
+from auth.middleware import user_data
+from config import setup_routes, settings
+
 
 asyncio.set_event_loop_policy(uvloop.EventLoopPolicy())
 logger = settings.log
 loop = asyncio.get_event_loop()
 
 async def get_app():
-    app = web.Application(debug=settings.DEBUG)
-    setup_routes(app)
-    aiohttp_jinja2.setup(app, loader=jinja2.FileSystemLoader(settings.TEMPLATES_DIR))
     redis_pool = await create_pool((settings.REDIS_HOST, settings.REDIS_PORT), db=settings.REDIS_DB)
-    setup(app, RedisStorage(redis_pool))
-    return app
+    web_app = web.Application(debug=settings.DEBUG, middlewares=[session_middleware(RedisStorage(redis_pool)),
+                                                                 user_data])
+    setup_routes(web_app)
+    aiohttp_jinja2.setup(web_app, loader=jinja2.FileSystemLoader(settings.TEMPLATES_DIR))
+    return web_app
 
 if __name__ == '__main__':
     app = loop.run_until_complete(get_app())
