+++ src/static/js/singup.js
