+++ src/templates/chat.html
@@ -1,5 +1,5 @@
 {% extends "base.html" %}
 {% block title %}Chat{% endblock %}
 {% block content %}
-    <h1>HELLO</h1>
+    <h1>hello, {{ user.first_name }}</h1>
 {% endblock %}
